requirements.txt
weaviate-client
openai
torch
